fix: change default type of ProgressInfo from float to double

+++ AutoMuteUs Portable.Shared/Entity/ProgressInfo/ProgressInfo.cs
@@ -1,6 +1,6 @@
 ﻿namespace AutoMuteUs_Portable.Shared.Entity.ProgressInfo;
 
-public class ProgressInfo : ProgressInfo<float>
+public class ProgressInfo : ProgressInfo<double>
 {
 }
